feat: move login to a separate pageэ
