+++ client/src/index.scss
@@ -23,11 +23,11 @@
 // 5. Components
 @import
 'style/components/button',
-'style/components/sign-in',
 'style/components/popup';
 
 // 6. Page-specific styles
 @import 
 'style/pages/home-page',
+'style/pages/login-page',
 'style/pages/not-found-page',
 'style/pages/plan-page';
