+++ client/src/ts/pages/loginPage/components/emailInput.ts
@@ -1,5 +1,5 @@
-import ErrorsList from '../base/enums/errorsList';
-import InputType from '../base/enums/inputTypes';
+import ErrorsList from '../../../base/enums/errorsList';
+import InputType from '../../../base/enums/inputTypes';
 import TextInput from './textInput';
 
 class EmailInput extends TextInput {
