+++ client/src/ts/pages/loginPage/components/secondPasswordInput.ts
@@ -1,5 +1,5 @@
-import ErrorsList from '../base/enums/errorsList';
-import InputType from '../base/enums/inputTypes';
+import ErrorsList from '../../../base/enums/errorsList';
+import InputType from '../../../base/enums/inputTypes';
 import FirstPasswordInput from './firstPasswordInput';
 import TextInput from './textInput';
