+++ client/src/ts/router.ts
@@ -1,5 +1,6 @@
 import RoutsList from './base/enums/routsList';
 import HomePage from './pages/homePage/homePage';
+import LoginPage from './pages/loginPage/loginPage';
 import NotFoundPage from './pages/notFoundPage/notFoundPage';
 import Page from './pages/page';
 import PlanPage from './pages/planPage/planPage';
@@ -10,6 +11,8 @@ class Router {
   static planPage: Page;
 
   static notFoundPage: Page;
+
+  static loginPage: Page;
 
   // static accountPage: Page;
 
@@ -26,6 +29,7 @@ class Router {
   constructor() {
     Router.homePage = new HomePage(this.goTo);
     Router.planPage = new PlanPage(this.goTo);
+    Router.loginPage = new LoginPage(this.goTo);
     Router.notFoundPage = new NotFoundPage(this.goTo);
   }
 
@@ -34,6 +38,9 @@ class Router {
     switch (pathname) {
       case RoutsList.planPage:
         Router.planPage.draw();
+        break;
+      case RoutsList.loginPage:
+        Router.loginPage.draw();
         break;
       case RoutsList.homePage:
         Router.homePage.draw();
